hyphen_alchemy_4/app/Data_purging_code/requirements.txt
configparser
logging
mysql-connector-python
psycopg2
python-dateutil